1. 修改回用全更新,用全更新是因为选择了item1之后,不止item1会改变,其他item也会改变,所以用全更新

+++ multilibrary/src/main/java/com/zhongjh/multimedia/album/widget/CheckView.kt
@@ -154,7 +154,7 @@ class CheckView : View {
             if (mChecked) {
                 initBackgroundPaint()
                 canvas.drawCircle(
-                    SIZE.toFloat() * mDensity / 2, SIZE.toFloat() * mDensity / 2,
+                    SIZE * mDensity / 2, SIZE * mDensity / 2,
                     BG_RADIUS * mDensity, mBackgroundPaint
                 )
                 mCheckDrawable!!.bounds = checkRect
